requirements.txt
azure-cognitiveservices-speech==1.7.0
azure-cognitiveservices-vision-face==0.4.0
azure-common==1.1.23
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
future==0.18.1
gTTS-token==1.1.3
idna==2.8
isodate==0.6.0
msrest==0.6.10
numpy==1.17.2
oauthlib==3.1.0
opencv-python==4.1.1.26
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.4
urllib3==1.25.6
